Fix for navigating between hub sites and footer disappearing
When you navigate across sites with the footer deployed, SharePoint downloads the page differently. The _bottomPlaceholder is returned as "isDisposed" - rather than when you refresh on a new page. When you then try and render the ReactDom it fails as _bottomPlaceholder is in a disposed state.
This now checks for isDisposed in the bottomPlaceholder